reusables-codes/user-auth-system/requirements.txt
fastapi[all]
uvicorn
python-jose[cryptography]
pwdlib[Argon2]
pydantic
pydantic_settings 
sqlalchemy